refactor: Moved components in my-v1-addon to my-v1-app
