--- docs/my-v1-addon/addon/components/products/product/card.css.d.ts
@@ -1,13 +1,0 @@
-declare const styles: {
-  readonly 'my-v1-addon-actions': string;
-  readonly 'my-v1-addon-body': string;
-  readonly 'my-v1-addon-container': string;
-  readonly 'my-v1-addon-description': string;
-  readonly 'my-v1-addon-header': string;
-  readonly 'my-v1-addon-image-container': string;
-  readonly 'my-v1-addon-link': string;
-  readonly 'my-v1-addon-name': string;
-  readonly 'my-v1-addon-price': string;
-};
-
-export default styles;
