--- docs/my-v1-addon/addon/components/products/product/image.css.d.ts
@@ -1,6 +1,0 @@
-declare const styles: {
-  readonly 'my-v1-addon-image': string;
-  readonly 'my-v1-addon-placeholder-image': string;
-};
-
-export default styles;
