+++ docs/my-v1-app/app/components/products/product/card.css.d.ts
@@ -1,0 +1,13 @@
+declare const styles: {
+  readonly 'actions': string;
+  readonly 'body': string;
+  readonly 'container': string;
+  readonly 'description': string;
+  readonly 'header': string;
+  readonly 'image-container': string;
+  readonly 'link': string;
+  readonly 'name': string;
+  readonly 'price': string;
+};
+
+export default styles;
