+++ docs/my-v1-app/app/components/products/product/card.ts
@@ -1,0 +1,21 @@
+import Component from '@glimmer/component';
+
+import type { Product } from '../../../data/products';
+import styles from './card.css';
+
+interface ProductsProductCardSignature {
+  Args: {
+    product: Product;
+    redirectTo?: string;
+  };
+}
+
+export default class ProductsProductCardComponent extends Component<ProductsProductCardSignature> {
+  styles = styles;
+}
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    'Products::Product::Card': typeof ProductsProductCardComponent;
+  }
+}
