+++ docs/my-v1-app/app/components/products/product/image.css
@@ -1,15 +1,15 @@
-.my-v1-addon-image,
-.my-v1-addon-placeholder-image {
+.image,
+.placeholder-image {
   aspect-ratio: 4 / 3;
   border-radius: 0.75rem;
   width: 100%;
 }
 
-.my-v1-addon-image {
+.image {
   object-fit: cover;
 }
 
-.my-v1-addon-placeholder-image {
+.placeholder-image {
   background: linear-gradient(
     36deg,
     rgb(255 224 130 / 40%) 15%,
