+++ docs/my-v1-app/app/components/products/product/image.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly 'image': string;
+  readonly 'placeholder-image': string;
+};
+
+export default styles;
