+++ docs/my-v1-app/app/components/products/product/image.hbs
