+++ docs/my-v1-app/app/components/products/product/image.ts
@@ -1,11 +1,7 @@
 import { isTesting, macroCondition } from '@embroider/macros';
 import Component from '@glimmer/component';
 
-/* TODO: Import styles from the stylesheet */
-const styles = {
-  image: 'my-v1-addon-image',
-  'placeholder-image': 'my-v1-addon-placeholder-image',
-};
+import styles from './image.css';
 
 interface ProductsProductImageSignature {
   Args: {
@@ -17,4 +13,10 @@ export default class ProductsProductImageComponent extends Component<ProductsPro
   isTestEnvironment = macroCondition(isTesting()) ? true : false;
 
   styles = styles;
+}
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    'Products::Product::Image': typeof ProductsProductImageComponent;
+  }
 }
