+++ docs/my-v1-app/tests/integration/components/products/product/image-test.ts
